backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
python-dotenv==0.19.0
werkzeug==3.1.3
pymupdf==1.20.0
spacy==3.8.7
pytesseract==0.3.13
Pillow==11.3.0
sentence-transformers
opencv
